refactor: Hide state in AddNewChartsWorkflow
